added login information alerts

+++ client/src/pages/Login/Login.js
@@ -8,6 +8,7 @@ class Login extends Component {
     state = {
         username: "",
         password: "",
+        usernameExist: false,
         loggedIn: false
     }
 
@@ -16,6 +17,24 @@ class Login extends Component {
         this.setState({
             [name]: value
         })
+
+        if (name === "username" && value) {
+            api.checkUsername(value)
+                .then(result => {
+                    if (result.data) {
+                        console.log("This username is found.");
+                        this.setState({
+                            usernameExist: true
+                        })
+                    } else {
+                        console.log("This username is not found.");
+                        this.setState({
+                            usernameExist: false
+                        })
+                    }
+                })
+                .catch(err => console.log(err));
+        }
     }
 
     handleFormSubmit = event => {
@@ -23,25 +42,70 @@ class Login extends Component {
 
         if (!this.state.username) {
             console.log("Please enter your username.");
+            let infoMissingText = document.createElement("p");
+            infoMissingText.className += "swal-warning-text";
+            infoMissingText.innerHTML = "Please enter your username.";
+            window.swal({
+                content: infoMissingText,
+                buttons: false,
+                icon: "warning",
+                timer: "1500"
+            });
         } else if (!this.state.password) {
             console.log("Please enter your password.");
+            let infoMissingText = document.createElement("p");
+            infoMissingText.className += "swal-warning-text";
+            infoMissingText.innerHTML = "Please enter your password.";
+            window.swal({
+                content: infoMissingText,
+                buttons: false,
+                icon: "warning",
+                timer: "1500"
+            });
+        } else if (this.state.username !== "" && this.state.usernameExist === false) {
+            console.log("User account does not exist.")
+            let infoWrongText = document.createElement("p");
+            infoWrongText.className += "swal-warning-text";
+            infoWrongText.innerHTML = "User account does not exist";
+            window.swal({
+                content: infoWrongText,
+                buttons: false,
+                icon: "warning",
+                timer: "1500"
+            });
         } else {
             api.login({
                 username: this.state.username,
                 password: this.state.password
             })
-            .then(loginResult => {
-                this.setState({ loggedIn: loginResult.data.isMatch });
-                this.props.loginUser(loginResult.data.isMatch);
-                
-                if (loginResult.data.isMatch) {
-                    console.log("Login successful.");
-                    sessionStorage.setItem("userId", loginResult.data.userId);
-                } else {
-                    console.log("Sorry, your login information was incorrect.");
-                }
-            })
-            .catch(err => console.log(err));
+                .then(loginResult => {
+                    this.setState({ loggedIn: loginResult.data.isMatch });
+                    this.props.loginUser(loginResult.data.isMatch);
+
+                    if (loginResult.data.isMatch) {
+                        console.log("Login successful.");
+                        let userLoggedInText = document.createElement("p");
+                        userLoggedInText.className += "swal-success-text animated tada";
+                        userLoggedInText.innerHTML = "Login successful.";
+                        window.swal({
+                            content: userLoggedInText,
+                            timer: 2000,
+                        });
+                        sessionStorage.setItem("userId", loginResult.data.userId);
+                    } else {
+                        console.log("Sorry, your login password was incorrect.");
+                        let infoWrongText = document.createElement("p");
+                        infoWrongText.className += "swal-warning-text";
+                        infoWrongText.innerHTML = "Sorry, your login password was incorrect.";
+                        window.swal({
+                            content: infoWrongText,
+                            buttons: false,
+                            icon: "warning",
+                            timer: "1500"
+                        });
+                    }
+                })
+                .catch(err => console.log(err));
         }
     }
 
@@ -77,7 +141,7 @@ class Login extends Component {
                         <button className="log-in-btn btn btn-lg btn-primary" onClick={this.handleFormSubmit}>Log In</button>
                     </form>
                     <p className="login-to-sign-up-text">Don't have an account? Sign up <Link to="/signup">here</Link>.</p>
-                    { this.state.loggedIn && <Redirect to="/" /> }
+                    {this.state.loggedIn && <Redirect to="/" />}
                 </div>
             </div>
         )
